+++ flang/runtime/extensions.cpp
@@ -10,10 +10,33 @@
 // extensions that will eventually be implemented in Fortran.
 
 #include "flang/Runtime/extensions.h"
+#include "terminator.h"
 #include "tools.h"
 #include "flang/Runtime/command.h"
 #include "flang/Runtime/descriptor.h"
 #include "flang/Runtime/io-api.h"
+#include <ctime>
+
+#ifdef _WIN32
+inline void CtimeBuffer(char *buffer, size_t bufsize, const time_t cur_time,
+    Fortran::runtime::Terminator terminator) {
+  int error{ctime_s(buffer, bufsize, &cur_time)};
+  RUNTIME_CHECK(terminator, error == 0);
+}
+#elif _POSIX_C_SOURCE >= 1 || _XOPEN_SOURCE || _BSD_SOURCE || _SVID_SOURCE || \
+    _POSIX_SOURCE
+inline void CtimeBuffer(char *buffer, size_t bufsize, const time_t cur_time,
+    Fortran::runtime::Terminator terminator) {
+  const char *res{ctime_r(&cur_time, buffer)};
+  RUNTIME_CHECK(terminator, res != nullptr);
+}
+#else
+inline void CtimeBuffer(char *buffer, size_t bufsize, const time_t cur_time,
+    Fortran::runtime::Terminator terminator) {
+  buffer[0] = '\0';
+  terminator.Crash("fdate is not supported.");
+}
+#endif
 
 #if _REENTRANT || _POSIX_C_SOURCE >= 199506L
 // System is posix-compliant and has getlogin_r
@@ -43,6 +66,26 @@ void FORTRAN_PROCEDURE_NAME(flush)(const int &unit) {
 }
 } // namespace io
 
+// CALL FDATE(DATE)
+void FORTRAN_PROCEDURE_NAME(fdate)(char *arg, std::int64_t length) {
+  // Day Mon dd hh:mm:ss yyyy\n\0 is 26 characters, e.g.
+  // Tue May 26 21:51:03 2015\n\0
+  char str[26];
+  // Insufficient space, fill with spaces and return.
+  if (length < 24) {
+    std::memset(arg, ' ', length);
+    return;
+  }
+
+  Terminator terminator{__FILE__, __LINE__};
+  std::time_t current_time;
+  std::time(&current_time);
+  CtimeBuffer(str, sizeof(str), current_time, terminator);
+
+  // Pad space on the last two byte `\n\0`, start at index 24 included.
+  CopyAndPad(arg, str, length, 24);
+}
+
 // RESULT = IARGC()
 std::int32_t FORTRAN_PROCEDURE_NAME(iargc)() { return RTNAME(ArgumentCount)(); }
 
@@ -57,13 +100,21 @@ void FORTRAN_PROCEDURE_NAME(getarg)(
 // CALL GETLOG(USRNAME)
 void FORTRAN_PROCEDURE_NAME(getlog)(std::byte *arg, std::int64_t length) {
 #if _REENTRANT || _POSIX_C_SOURCE >= 199506L
-  const int nameMaxLen{LOGIN_NAME_MAX + 1};
-  char str[nameMaxLen];
+  int nameMaxLen;
+#ifdef LOGIN_NAME_MAX
+  nameMaxLen = LOGIN_NAME_MAX + 1;
+#else
+  nameMaxLen = sysconf(_SC_LOGIN_NAME_MAX) + 1;
+  if (nameMaxLen == -1)
+    nameMaxLen = _POSIX_LOGIN_NAME_MAX + 1;
+#endif
+  std::vector<char> str(nameMaxLen);
 
-  int error{getlogin_r(str, nameMaxLen)};
+  int error{getlogin_r(str.data(), nameMaxLen)};
   if (error == 0) {
     // no error: find first \0 in string then pad from there
-    CopyAndPad(reinterpret_cast<char *>(arg), str, length, std::strlen(str));
+    CopyAndPad(reinterpret_cast<char *>(arg), str.data(), length,
+        std::strlen(str.data()));
   } else {
     // error occur: get username from environment variable
     GetUsernameEnvVar("LOGNAME", arg, length);