+++ llvm/docs/ReleaseNotes.rst
@@ -73,6 +73,8 @@ Changes to the LLVM IR
 Changes to LLVM infrastructure
 ------------------------------
 
+* Minimum Clang version to build LLVM in C++20 configuration has been updated to clang-17.0.6.
+
 Changes to building LLVM
 ------------------------
 
@@ -151,6 +153,7 @@ Changes to the RISC-V Backend
   "SiFive Custom Instruction Extension" as SiFive defines it. The LLVM project
   needs to work with SiFive to define and document real extension names for
   individual CSRs and instructions.
+* ``-mcpu=sifive-p450`` was added.
 
 Changes to the WebAssembly Backend
 ----------------------------------